+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 
 ## 0.0.2 - under development
 
+### Added
 - improve release script
 - add flags for ratio and number of requests
 - add position argument with urls
+
+### Changed
+- get.go to requests.go and refactor
 
 ## 0.0.1 - 2019-10-04
 ### Added